cleaned up code started packets

+++ rpgNecroPaladin/items/ItemGrandSword.java
@@ -2,7 +2,6 @@ package rpgNecroPaladin.items;
 
 import net.minecraft.block.Block;
 import net.minecraft.creativetab.CreativeTabs;
-import net.minecraft.entity.EntityLiving;
 import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.init.Blocks;
@@ -54,8 +53,9 @@ public class ItemGrandSword extends Item {
 	@Override
 	public boolean getIsRepairable(ItemStack par1ItemStack,
 			ItemStack par2ItemStack) {
-		return this.toolMaterial.getToolCraftingMaterial() == par2ItemStack.getItem() ? true
-				: super.getIsRepairable(par1ItemStack, par2ItemStack);
+		return this.toolMaterial.getToolCraftingMaterial() == par2ItemStack
+				.getItem() ? true : super.getIsRepairable(par1ItemStack,
+				par2ItemStack);
 	}
 
 	/**
@@ -92,10 +92,9 @@ public class ItemGrandSword extends Item {
 
 	@Override
 	public boolean onBlockDestroyed(ItemStack par1ItemStack, World par2World,
-			Block b , int par4, int par5, int par6,
+			Block b, int par4, int par5, int par6,
 			EntityLivingBase par7EntityLiving) {
-		if ((double) b.getBlockHardness(par2World, par4,
-				par5, par6) != 0.0D) {
+		if (b.getBlockHardness(par2World, par4, par5, par6) != 0.0D) {
 			par1ItemStack.damageItem(2, par7EntityLiving);
 		}
 		return true;
